requirements.txt
Click==7.0
Flask==1.1.1
Flask-API==2.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.10.0
Werkzeug==0.16.0